+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Improve highlight detection logic in DOCX ➡️ Markdown converter
 - Add option to preserve anchored and floating images in DOCX ➡️ HTML/Markdown converter
 - Fix: inline images incorrectly always created a new line in DOCX ➡️ HTML converter
-- Preserve hyperlinks on images in DOCX ➡️ HTML/RTF converter
+- Preserve hyperlinks on images in DOCX ➡️ HTML/RTF/PDF converter
 - Fix: hyperlinks and images in header/footer/footnote/endnote were sometimes lost in DOCX converters
 - Preserve image alternate text in DOCX ➡️ HTML/Markdown converter
 - The new `WriteImageDescription` property allows to write alternate text / description (if available) in place of images in DOCX ➡️ plain text converter (*false* by default)